pyproject.toml
[project]
name = "cloudstorage2bq"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "craig", email = "craiig@gmail.com" }
]
dependencies = [
    "google-cloud-bigquery>=3.18.0",
    "google-cloud-storage>=2.15.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
cloudstorage2bq = "cloudstorage2bq.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/cloudstorage2bq"]
